Consolidate migrations: squash alter-tables and remove data-fix migrations
Absorb incremental schema changes into their parent create-table migrations
(bed_types, fee_types, bath_room_types, charge_bases, system_settings,
calendar_days) and remove 7 data-fix migrations whose logic is already
captured in the current seeders. Reduces 36 migrations to 21.

+++ database/migrations/2026_03_23_220108_create_calendar_days_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->date('holiday_original_date')->nullable();
             $table->date('holiday_observed_date')->nullable();
             $table->string('holiday_group', 20)->nullable();
-            $table->decimal('holiday_impact', 4, 1)->nullable();
+            $table->unsignedTinyInteger('holiday_impact')->nullable();
             $table->boolean('is_bridge_day')->default(false);
             $table->foreignId('season_block_id')->nullable()->constrained()->nullOnDelete();
             $table->string('season_block_name', 100)->nullable();
